database: Split models.go into different files each contains one model
